🚚 Move emoji picker to own namespace folder
