+++ PuppetBotClient/PuppetBotClient/Views/EmojiPicker/ServerEmojiCollectionView.axaml
@@ -2,10 +2,10 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:views="clr-namespace:PuppetBotClient.Views"
+             xmlns:views="clr-namespace:PuppetBotClient.Views.EmojiPicker"
              mc:Ignorable="d" d:DesignWidth="450" d:DesignHeight="450"
              Width="400" Height="420"
-             x:Class="PuppetBotClient.Views.ServerEmojiCollectionView">
+             x:Class="PuppetBotClient.Views.EmojiPicker.ServerEmojiCollectionView">
   <Canvas>
     <Label Name="ServerNameLabel" Width="400" Height="30" FontSize="15" FontWeight="Bold">Server Name</Label>
     <StackPanel Canvas.Top="30" Name="EmojiCollectionCanvas" Orientation="Vertical" Height="50" Width="350">
