+++ PuppetBotClient/PuppetBotClient/Views/SetStatusDialog.axaml
@@ -1,0 +1,9 @@
+<Window xmlns="https://github.com/avaloniaui"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+        x:Class="PuppetBotClient.Views.SetStatusDialog"
+        Title="SetStatusDialog">
+  Welcome to Avalonia!
+</Window>
